+++ src/testing/index.ts
@@ -1,2 +1,10 @@
 export { describeTransform, snapshotTransform, toMatchShape } from "./helpers.js";
-export { runStandardTests } from "./run.js";
+export {
+	runStandardTests,
+	type StandardTestsManifest,
+	type StandardTestsOptions,
+	type StandardTestsResult,
+	type StandardTestsUpstreamCall,
+	type StandardTestsUpstreamResponse,
+	type StandardTestsUpstreamStub,
+} from "./run.js";